add policies

+++ src/controllers/prompts_controller.ts
@@ -1,9 +1,15 @@
 import { Request, Response } from 'express'
 
-import { PromptInterface, UserModel } from '@src/models'
-import { render200, render404, serverError } from '@src/utils/serverErrors'
+import { PromptInterface, PromptModel, UserModel } from '@src/models'
+import { PromptsPolicy } from '@src/policies'
+import { DEVELOPER_EMAIL, FREE_USER_PROMPTS_LIMIT } from '@src/utils/constants'
+import { render200, render404, renderUnuathorized, serverError } from '@src/utils/serverErrors'
 
 import PromptsRepository from '../repositories/prompts.repository'
+
+const findPrompt = ({ userId, promptId }: { userId: number; promptId: number }): Promise<PromptModel> => {
+  return PromptsRepository.findBy({ user_id: userId, id: promptId })
+}
 
 export const show = async (req: Request, res: Response): Promise<void> => {
   const currentUser: UserModel = req.user as UserModel
@@ -36,12 +42,20 @@ export const all = async (req: Request, res: Response): Promise<void> => {
 export const create = async (req: Request, res: Response): Promise<void> => {
   const currentUser: UserModel = req.user as UserModel
   try {
-    const params = {
-      ...req.body,
-      user_id: currentUser.id,
-    } as PromptInterface
-    const newPrompt = await PromptsRepository.create(params)
-    res.status(200).json(newPrompt)
+    const canCreate = await new PromptsPolicy(currentUser).create()
+    if (canCreate) {
+      const params = {
+        ...req.body,
+        user_id: currentUser.id,
+      } as PromptInterface
+      const newPrompt = await PromptsRepository.create(params)
+      res.status(200).json(newPrompt)
+    } else {
+      renderUnuathorized({
+        res,
+        message: `On a free account you can create only ${FREE_USER_PROMPTS_LIMIT} prompts. If you're excited about promptinator and would like to unlock more features <a href="mailto:${DEVELOPER_EMAIL}">contact us</a>`,
+      })
+    }
   } catch (err) {
     serverError({ res, exception: err as Error })
   }
@@ -50,21 +64,32 @@ export const create = async (req: Request, res: Response): Promise<void> => {
 export const update = async (req: Request, res: Response): Promise<void> => {
   const currentUser: UserModel = req.user as UserModel
   try {
-    const params = {
-      ...req.body,
-      user_id: currentUser.id,
-    } as PromptInterface
-    const prompt = await PromptsRepository.update(req.params.id, params)
-    res.status(200).json(prompt)
+    let prompt: PromptModel = await findPrompt({ userId: currentUser.id as number, promptId: Number(req.params.id) })
+    if (await new PromptsPolicy(currentUser, prompt).create()) {
+      const params = {
+        ...req.body,
+        user_id: currentUser.id,
+      } as PromptInterface
+      prompt = await PromptsRepository.update(req.params.id, params)
+      res.status(200).json(prompt)
+    } else {
+      render404({ message: 'Prompt not found', res })
+    }
   } catch (err) {
     serverError({ res, exception: err as Error })
   }
 }
 
 export const deletePrompt = async (req: Request, res: Response): Promise<void> => {
+  const currentUser: UserModel = req.user as UserModel
   try {
-    await PromptsRepository.delete(req.params.id)
-    render200({ res, message: 'Prompt deleted' })
+    const prompt: PromptModel = await findPrompt({ userId: currentUser.id as number, promptId: Number(req.params.id) })
+    if (await new PromptsPolicy(currentUser, prompt).delete()) {
+      await PromptsRepository.delete(req.params.id)
+      render200({ res, message: 'Prompt deleted' })
+    } else {
+      render404({ message: 'Prompt not found', res })
+    }
   } catch (err) {
     serverError({ res, exception: err as Error })
   }
